move classification out as a separate plugin
