+++ src/js/plugins/classification/pre-publish-classify-post.js
